+++ odroid/energymon-odroid.h
@@ -11,18 +11,19 @@
 extern "C" {
 #endif
 
+#include <inttypes.h>
 #include <stddef.h>
 #include "energymon.h"
 
 int energymon_init_odroid(energymon* em);
 
-unsigned long long energymon_read_total_odroid(const energymon* em);
+uint64_t energymon_read_total_odroid(const energymon* em);
 
 int energymon_finish_odroid(energymon* em);
 
 char* energymon_get_source_odroid(char* buffer, size_t n);
 
-unsigned long long energymon_get_interval_odroid(const energymon* em);
+uint64_t energymon_get_interval_odroid(const energymon* em);
 
 int energymon_get_odroid(energymon* impl);
